Defend against duplicate entries.
You WILL see duplicate OnMailReady events if something else is iterating the mail list at the same time we are. I've seen this enough to see "36 messages saved" when we only had 8 actual messages in the mailbox.

Doesn't quite work, the iteration to flatten the key-by-mail_id dict back to an array isn't seeing anything.